Introduce getTokenAddress
Currently, we have fork network and inject fork network in
tokens.config.ts. We also use DeploymentOrAddress for tokens that we
own. However, we do not yet have a home for external token. For example,
there are lots of duplication for address of RT1. This is error-prone
and it makes sense to consolidate all token address in one conifg.

Testing done:
> FORK_NETWORK=arbitrum_mainnet hh test test/config/TokenAddress.spec.ts
>> Error: No config found for token WOM in network arbitrum_mainnet

> hh test test/config/TokenAddress.spec.ts
>> passes