Fix ethers provider when connecting to mainnet
The workaround to allow impersonation is triggered whenever a url is
specified. This includes localhost and mainnet. When the default
provider is overriden, it does not have the signers from hardhat config,
since this functionality relies on hardhat's middleware. Therefore, a
"unknown account" error will be thrown. This is reproducible by send
transaction. Note: deployment would still work, since the deploy
function of hardhat-deploy has internal logic to look up the correct
signers (whose private key we supplied via hardhat config).

This will repro, when script trying to send approvals txn
> hh deploy --tags WombatRouter --network bsc_testnet

This will not repro, when we just do smart contract deployment
> hh deploy --tags WombatToken --network bsc_testnet